+++ linkerd/http/retry/src/replay/buffer.rs
@@ -1,0 +1,151 @@
+use bytes::{Buf, BufMut, Bytes, BytesMut};
+use std::{collections::VecDeque, io::IoSlice};
+
+/// Data returned by `ReplayBody`'s `http_body::Body` implementation is either
+/// `Bytes` returned by the initial body, or a list of all `Bytes` chunks
+/// returned by the initial body (when replaying it).
+#[derive(Debug)]
+pub enum Data {
+    Initial(Bytes),
+    Replay(BufList),
+}
+
+/// Body data composed of multiple `Bytes` chunks.
+#[derive(Clone, Debug, Default)]
+pub struct BufList {
+    bufs: VecDeque<Bytes>,
+}
+
+// === impl Data ===
+
+impl Buf for Data {
+    #[inline]
+    fn remaining(&self) -> usize {
+        match self {
+            Data::Initial(buf) => buf.remaining(),
+            Data::Replay(bufs) => bufs.remaining(),
+        }
+    }
+
+    #[inline]
+    fn chunk(&self) -> &[u8] {
+        match self {
+            Data::Initial(buf) => buf.chunk(),
+            Data::Replay(bufs) => bufs.chunk(),
+        }
+    }
+
+    #[inline]
+    fn chunks_vectored<'iovs>(&'iovs self, iovs: &mut [IoSlice<'iovs>]) -> usize {
+        match self {
+            Data::Initial(buf) => buf.chunks_vectored(iovs),
+            Data::Replay(bufs) => bufs.chunks_vectored(iovs),
+        }
+    }
+
+    #[inline]
+    fn advance(&mut self, amt: usize) {
+        match self {
+            Data::Initial(buf) => buf.advance(amt),
+            Data::Replay(bufs) => bufs.advance(amt),
+        }
+    }
+
+    #[inline]
+    fn copy_to_bytes(&mut self, len: usize) -> Bytes {
+        match self {
+            Data::Initial(buf) => buf.copy_to_bytes(len),
+            Data::Replay(bufs) => bufs.copy_to_bytes(len),
+        }
+    }
+}
+
+// === impl BufList ===
+
+impl BufList {
+    pub(super) fn push_chunk(&mut self, mut data: impl Buf) -> Bytes {
+        let len = data.remaining();
+        // `data` is (almost) certainly a `Bytes`, so `copy_to_bytes` should
+        // internally be a cheap refcount bump almost all of the time.
+        // But, if it isn't, this will copy it to a `Bytes` that we can
+        // now clone.
+        let bytes = data.copy_to_bytes(len);
+        // Buffer a clone of the bytes read on this poll.
+        self.bufs.push_back(bytes.clone());
+        // Return the bytes
+        bytes
+    }
+}
+
+impl Buf for BufList {
+    fn remaining(&self) -> usize {
+        self.bufs.iter().map(Buf::remaining).sum()
+    }
+
+    fn chunk(&self) -> &[u8] {
+        self.bufs.front().map(Buf::chunk).unwrap_or(&[])
+    }
+
+    fn chunks_vectored<'iovs>(&'iovs self, iovs: &mut [IoSlice<'iovs>]) -> usize {
+        // Are there more than zero iovecs to write to?
+        if iovs.is_empty() {
+            return 0;
+        }
+
+        // Loop over the buffers in the replay buffer list, and try to fill as
+        // many iovecs as we can from each buffer.
+        let mut filled = 0;
+        for buf in &self.bufs {
+            filled += buf.chunks_vectored(&mut iovs[filled..]);
+            if filled == iovs.len() {
+                return filled;
+            }
+        }
+
+        filled
+    }
+
+    fn advance(&mut self, mut amt: usize) {
+        while amt > 0 {
+            let rem = self.bufs[0].remaining();
+            // If the amount to advance by is less than the first buffer in
+            // the buffer list, advance that buffer's cursor by `amt`,
+            // and we're done.
+            if rem > amt {
+                self.bufs[0].advance(amt);
+                return;
+            }
+
+            // Otherwise, advance the first buffer to its end, and
+            // continue.
+            self.bufs[0].advance(rem);
+            amt -= rem;
+
+            self.bufs.pop_front();
+        }
+    }
+
+    fn copy_to_bytes(&mut self, len: usize) -> Bytes {
+        // If the length of the requested `Bytes` is <= the length of the front
+        // buffer, we can just use its `copy_to_bytes` implementation (which is
+        // just a reference count bump).
+        match self.bufs.front_mut() {
+            Some(first) if len <= first.remaining() => {
+                let buf = first.copy_to_bytes(len);
+                // If we consumed the first buffer, also advance our "cursor" by
+                // popping it.
+                if first.remaining() == 0 {
+                    self.bufs.pop_front();
+                }
+
+                buf
+            }
+            _ => {
+                assert!(len <= self.remaining(), "`len` greater than remaining");
+                let mut buf = BytesMut::with_capacity(len);
+                buf.put(self.take(len));
+                buf.freeze()
+            }
+        }
+    }
+}
